+++ src/main/java/de/tinf22b6/dhbwhub/model/log_tables/LikeLogtableEventComment.java
@@ -1,6 +1,6 @@
-package de.tinf22b6.dhbwhub.model.logtables;
+package de.tinf22b6.dhbwhub.model.log_tables;
 
-import de.tinf22b6.dhbwhub.model.Comment;
+import de.tinf22b6.dhbwhub.model.EventComment;
 import de.tinf22b6.dhbwhub.model.User;
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -13,11 +13,11 @@ import lombok.Setter;
 @RequiredArgsConstructor
 @NoArgsConstructor(force = true)
 @Entity
-@Table(name = "like_logtable_post_comment")
-public class LikeLogtablePostComment {
+@Table(name = "like_logtable_event_comment")
+public class LikeLogtableEventComment {
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "log_comment_generator")
-    @SequenceGenerator(name = "log_comment_generator", sequenceName = "log_comment_seq", allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "log_event_comment_generator")
+    @SequenceGenerator(name = "log_event_comment_generator", sequenceName = "log_event_comment_seq", allocationSize = 1)
     @Column(name = "log_id")
     private Long id;
 
@@ -25,7 +25,7 @@ public class LikeLogtablePostComment {
     @ManyToOne(cascade = CascadeType.PERSIST)
     private final User user;
 
-    @JoinColumn(name = "comment_id")
+    @JoinColumn(name = "event_comment_id")
     @ManyToOne(cascade = CascadeType.PERSIST)
-    private final Comment comment;
+    private final EventComment eventComment;
 }
