+++ src/main/java/de/tinf22b6/dhbwhub/repository/interfaces/SpringLogtableEventPostRepository.java
@@ -1,6 +1,6 @@
 package de.tinf22b6.dhbwhub.repository.interfaces;
 
-import de.tinf22b6.dhbwhub.model.logtables.LikeLogtableEventPost;
+import de.tinf22b6.dhbwhub.model.log_tables.LikeLogtableEventPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
