+++ src/main/java/de/tinf22b6/dhbwhub/repository/interfaces/SpringLogtablePostRepository.java
@@ -1,6 +1,6 @@
 package de.tinf22b6.dhbwhub.repository.interfaces;
 
-import de.tinf22b6.dhbwhub.model.logtables.LikeLogtablePost;
+import de.tinf22b6.dhbwhub.model.log_tables.LikeLogtablePost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
